royalnova/zkp_privacy.py
# zkp_privacy.py
# Royalnova (RNX) Zero-Knowledge Proof Privacy
from py_ecc.bn128 import G1, G2, add, multiply, pairing, neg, eq
import hashlib
import random

class ZKPPrivacy:
    def __init__(self):
        # Initialize with random private key
        self.private_key = self.generate_private_key()
        self.g1 = G1
        self.g2 = G2

        # Trusted setup parameters
        self.pk, self.vk = self.trusted_setup()

    def generate_private_key(self):
        """Generate random private key"""
        return random.randint(1, 2**256)

    def trusted_setup(self):
        """ZK-SNARK setup phase (trusted setup)"""
        pk = multiply(self.g1, self.private_key)
        vk = multiply(self.g2, self.private_key)
        return pk, vk

    def generate_proof(self, message):
        """Generate ZKP proof"""
        message_hash = hashlib.sha256(message.encode()).digest()
        proof = multiply(G1, int.from_bytes(message_hash, 'big') + self.private_key)
        return proof

    def verify_proof(self, proof, message):
        """Verify ZKP proof"""
        message_hash = hashlib.sha256(message.encode()).digest()

        # Pairing checks
        left = pairing(proof, self.g2)
        right = pairing(multiply(self.g1, int.from_bytes(message_hash, 'big')), self.vk)
        
        return eq(left, right)

# Simulate ZKP process
if __name__ == "__main__":
    print("\n🔒 Royalnova (RNX) ZKP Privacy Demo 🔒")

    zkp = ZKPPrivacy()

    # Generate proof for a transaction message
    message = "Send 100 RNX from Alice to Bob"
    proof = zkp.generate_proof(message)
    print("\n✅ Proof Generated")

    # Verify proof
    is_valid = zkp.verify_proof(proof, message)

    if is_valid:
        print("\n🎯 ZKP Verified: Transaction is valid and private!")
    else:
        print("\n❌ ZKP Verification Failed")

